requirements.txt
numpy
scipy
cython
h5py
Pillow
scikit-image
tensorflow
keras
jupyter
pyserial
